+++ package.json
@@ -34,6 +34,7 @@
     "mochawesome-report-generator": "^5.1.0",
     "node-sass": "^4.14.1",
     "parcel": "^1.12.4",
+    "sass": "^1.68.0",
     "stylelint": "^13.5.0",
     "stylelint-config-recommended-scss": "^3.3.0",
     "stylelint-scss": "^3.17.2"
