--- src/scripts/constants.js
@@ -1,14 +1,0 @@
-export const BUTTON_CLASS = 'button';
-export const FIELD_ROW_CLASS = 'field-row';
-export const FIELD_CELL_CLASS = 'field-cell';
-export const GAME_SCORE_CLASS = 'game-score';
-export const START_MESSAGE_CLASS = 'message-start';
-export const LOSE_MESSAGE_CLASS = 'message-lose';
-export const WIN_MESSAGE_CLASS = 'message-win';
-export const HIDDEN_CLASS = 'hidden';
-
-export const RESTART_TEXT = 'Restart';
-export const START_TEXT = 'Start';
-
-export const FIELD_SIZE = 4;
-export const WIN_SCORE = 2048;
